[minor][table-planner-blink] Move CommonPhysicalJoin and CommonPhysicalLookupJoin to `nodes.physical.common` package
This closes #14733
