+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/plan/nodes/logical/FlinkLogicalLegacySink.scala
@@ -44,7 +44,7 @@ class FlinkLogicalLegacySink(
     val catalogTable: CatalogTable,
     val staticPartitions: Map[String, String])
   extends LegacySink(cluster, traitSet, input, sink, sinkName)
-          with FlinkLogicalRel {
+  with FlinkLogicalRel {
 
   override def copy(traitSet: RelTraitSet, inputs: util.List[RelNode]): RelNode = {
     new FlinkLogicalLegacySink(
